assistants/explorer-doc-edit-assistant/pyproject.toml
[project]
name = "explorer-doc-edit-assistant"
version = "0.1.0"
description = "A Python Semantic Workbench assistant for exploring doc edit capabilities."
authors = [{ name = "Semantic Workbench Team" }]
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "anthropic>=0.40.0",
    "anthropic-client>=0.1.0",
    "assistant-drive>=0.1.0",
    "assistant-extensions[attachments]>=0.1.0",
    "content-safety>=0.1.0",
    "deepmerge>=2.0",
    "html2docx>=1.6.0",
    "markdown>=3.6",
    "openai>=1.61",
    "openai-client>=0.1.0",
    "pendulum>=3.0",
    "python-liquid>=1.13",
]

[tool.hatch.build.targets.wheel]
packages = ["assistant"]

[tool.uv]
package = true

[tool.uv.sources]
anthropic-client = { path = "../../libraries/python/anthropic-client", editable = true }
assistant-drive = { path = "../../libraries/python/assistant-drive", editable = true }
assistant-extensions = { path = "../../libraries/python/assistant-extensions", editable = true }
content-safety = { path = "../../libraries/python/content-safety/", editable = true }
openai-client = { path = "../../libraries/python/openai-client", editable = true }

[tool.ruff]
line-length = 120
target-version = "py311"

[tool.ruff.lint]
unfixable = ["F401"]

[tool.ruff.format]
docstring-code-format = true
line-ending = "lf"
preview = true

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[dependency-groups]
dev = ["pyright>=1.1.389", "ruff"]
